Promise Update for Model

+++ model/admin/admin.js
@@ -3,27 +3,23 @@ const connectionPool = require('../../conf/db');
 exports.admin_create = (admin_info) => {
     return new Promise((resolve, reject) => {
         connectionPool.getConnection((connectionError, connection) => {
-            if (connectionError) {
-                reject('數據庫連接失敗');
-            } else {
-                connection.query(`SELECT * FROM companydb.admin
-                                  WHERE admin_email = ?`, [admin_info.admin_email], (error, result) => {
-                    if (error) {
-                        reject('數據庫連接失敗');
-                    } else if (result.length > 0) {
-                        reject('該郵箱已存在，請使用新的郵箱注冊');
-                    } else {
-                        connection.query('INSERT INTO companydb.admin SET ?', admin_info, (error, result) => {
-                            if (error) {
-                                reject('數據庫連接失敗');
-                            } else if (result.affectedRows === 1) {
-                                resolve(`${admin_info.admin_name} 注冊成功`);
-                            }
-                        });
-                    }
+            if (connectionError) reject(`數據庫連接失敗`);
+            return connection.promise().query(`SELECT * FROM companydb.admin WHERE admin_email = ? AND admin_role = 1`, [admin_info.admin_email])
+                .then(([rows, field]) => {
+                    if (rows.length) reject(`該郵箱已存在，請使用新的郵箱注冊`);
+                    else return connection.promise().query(`INSERT INTO companydb.admin SET ?`, [admin_info])
                 })
-            }
-            connection.release();
+                .then((result) => {
+                    if (result[0].affectedRows === 1) resolve(`${admin_info.admin_name} 注冊成功`);
+                    else reject(`資料新增失敗`);
+                })
+                .catch((err) => {
+                    console.error(`CATCH ERROR : ${err}`);
+                    reject(`系統暫時無法運行該功能`);
+                })
+                .finally(() => {
+                    connection.release();
+                })
         });
     });
 };
@@ -31,21 +27,18 @@ exports.admin_create = (admin_info) => {
 exports.admin_display = (admin_id) => {
     return new Promise((resolve, reject) => {
         connectionPool.getConnection((connectionError, connection) => {
-            if (connectionError) {
-                reject(connectionError);
-            } else {
-                connection.query(`SELECT * FROM companydb.admin
-                                  WHERE admin_id = ?`, [admin_id], (error, result) => {
-                    if (error) {
-                        reject('系統暫時無法運行該功能');
-                    } else if (result.length > 0) {
-                        resolve(result)
-                    } else {
-                        resolve(result);
-                    }
+            if (connectionError) reject(`數據庫連接失敗`);
+            return connection.promise().query(`SELECT * FROM companydb.admin WHERE admin_id = ?`, [admin_id])
+                .then(([rows, field]) => {
+                    resolve(rows);
+                })
+                .catch((err) => {
+                    console.error(`CATCH ERROR : ${err}`);
+                    reject(`系統暫時無法運行該功能`);
+                })
+                .finally(() => {
                     connection.release();
-                });
-            }
+                })
         });
     });
 };
@@ -53,23 +46,18 @@ exports.admin_display = (admin_id) => {
 exports.admin_display_list = () => {
     return new Promise((resolve, reject) => {
         connectionPool.getConnection((connectionError, connection) => {
-            if (connectionError) {
-                reject(connectionError);
-            } else {
-                connection.query(`SELECT admin_id, admin_name,
-                                  DATE_FORMAT(last_login, '%W %M %Y %H:%i:%s') AS last_login 
-                                  FROM companydb.admin
-                                  WHERE admin_role = 1`, (error, result) => {
-                    if (error) {
-                        reject('系統暫時無法運行該功能');
-                    } else if (result.length > 0) {
-                        resolve(result)
-                    } else {
-                        resolve(result);
-                    }
+            if (connectionError) reject(`數據庫連接失敗`);
+            return connection.promise().query(`SELECT admin_id, admin_name, DATE_FORMAT(last_login, '%W %M %Y %H:%i:%s') AS last_login FROM companydb.admin WHERE admin_role = 1`)
+                .then(([rows, field]) => {
+                    resolve(rows)
+                })
+                .catch((err) => {
+                    console.error(`CATCH ERROR : ${err}`);
+                    reject(`系統暫時無法運行該功能`);
+                })
+                .finally(() => {
                     connection.release();
-                });
-            }
+                })
         });
     });
 };
@@ -77,25 +65,22 @@ exports.admin_display_list = () => {
 exports.admin_update = (admin_id, admin_info) => {
     return new Promise((resolve, reject) => {
         connectionPool.getConnection((connectionError, connection) => {
-            if (connectionError) {
-                reject(connectionError);
-            } else {
-                connection.query(`UPDATE companydb.admin SET ?
-                                  WHERE admin_id = ?`, [admin_info, admin_id], (error, result) => {
-                    if (error) {
-                        console.error('SQL Error : ', error);
-                        reject('數據庫連接失敗');
-                    } else if (result.affectedRows === 0) {
-                        resolve('資料更新失敗')
-                    } else if (result.message.match('Changed: 1')) {
-                        resolve('資料更新成功')
-                    }
-                    else {
-                        resolve('資料無異動');
-                    }
-                    connection.release();
-                });
-            }
+            connectionPool.getConnection((connectionError, connection) => {
+                if (connectionError) reject(`數據庫連接失敗`);
+                return connection.promise().query(`UPDATE companydb.admin SET ? WHERE admin_id = ?`, [admin_info, admin_id])
+                    .then((result) => {
+                        if (result[0].affectedRows === 0) reject(`資料更新失敗`);
+                        else if (result[0].info.match('Changed: 1')) resolve(`${admin_info.admin_name} 資料更新成功`);
+                        else resolve(`資料沒有異動`);
+                    })
+                    .catch((err) => {
+                        console.error(`CATCH ERROR : ${err}`);
+                        reject(`資料更新失敗`);
+                    })
+                    .finally(() => {
+                        connection.release();
+                    })
+            });
         });
     });
 };
@@ -103,21 +88,19 @@ exports.admin_update = (admin_id, admin_info) => {
 exports.admin_delete = (admin_id) => {
     return new Promise((resolve, reject) => {
         connectionPool.getConnection((connectionError, connection) => {
-            if (connectionError) {
-                reject(connectionError);
-            } else {
-                connection.query(`DELETE FROM companydb.admin
-                                  WHERE admin_id = ?`, [admin_id], (error, result) => {
-                    if (error) {
-                        reject('數據庫連接失敗');
-                    } else if (result.affectedRows === 1) {
-                        resolve('資料刪除成功')
-                    } else {
-                        reject('資料刪除失敗');
-                    }
+            if (connectionError) reject(`數據庫連接失敗`);
+            return connection.promise().query(`DELETE FROM companydb.admin WHERE admin_id = ?`, [admin_id])
+                .then((result) => {
+                    if (result[0].affectedRows === 1) resolve(`資料刪除成功`);
+                    else reject(`資料刪除失敗`);
+                })
+                .catch((err) => {
+                    console.error(`SQL ERROR : ${err}`);
+                    reject(`資料刪除失敗`);
+                })
+                .finally(() => {
                     connection.release();
-                });
-            }
+                })
         });
     });
 };
